refactor : 주문 리팩토링 외 사용 하지 않는 클래스 삭제

+++ src/test/java/kitchenpos/common/DatabaseCleanup.java
@@ -34,6 +34,9 @@ public class DatabaseCleanup implements InitializingBean {
         entityManager.createNativeQuery("SET REFERENTIAL_INTEGRITY FALSE").executeUpdate();
 
         for (String tableName : tableNames) {
+            if ("order".equals(tableName)) {
+                continue;
+            }
             entityManager.createNativeQuery("TRUNCATE TABLE " + tableName).executeUpdate();
         }
